Added refresh button to info tab and refresh action pane immediately on new action
- Added refresh button to document Info tab: refreshes all tabs.
- Moved updateDocumentActions function to DocumentActionsContext to make it available across application.
- Refresh document actions when new action added.
